requirements.txt
openai>=1.40.0
pandas
streamlit
python-dotenv
